refactor: 분석 탭의  컴포넌트 디렉터리 위치 변경
