+++ src/frontend/src/router/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import VueRouter from 'vue-router';
 import Feed from '../views/Feed.vue';
 import Post from '../views/Post.vue';
-import Analysis from '../views/Analysis.vue';
+import Analysis from '../views/Analysis/Analysis.vue';
 import Diary from '../views/Diary/Diary.vue';
 
 import Header from '@/components/Header';
