+++ src/frontend/src/views/Analysis/Analysis.vue
@@ -34,7 +34,7 @@
 
 <script>
 import MyPageTabs from '@/components/MyPageTabs.vue';
-import RadarChart from '@/components/RadarChart.vue';
+import RadarChart from './components/RadarChart.vue';
 export default {
   data() {
     return {};
