+++ src/frontend/src/views/Analysis/components/RadarChart.vue
